Move Player observer to data
